remove the incStage logic
seems to be too risky, if something happens and the incStage method
fails, due to missing a deadline the whole ICO goes down the drain. It
costs a bit more gas to automatically get the stage index but I think it
is worth it due to minimizing the failure area.